src/backend/requirements.txt
fastapi == 0.95.0

uvicorn == 0.21.1

motor == 3.1.1

python_dotenv == 1.0.0
